Refactor to extract dummy events into the shared support folder
